Move searchbox to top even if no results or error.
When there are no results or an error the message
was shown at the bottom of the screen and it was
very hard to see. This lead to the user to think that
the app is broken because the message might not be
visible.

Now the searchbox moves to the top in those cases
so the message is visible.